working individual post component to simplify code
