+++ src/components/Pages/IndividualPost/FullIndividualPost.module.css
@@ -1,0 +1,4 @@
+.postContainer {
+  display: flex;
+  justify-content: center;
+}
