+++ src/components/Pages/IndividualPost/IndividualPost.module.css
@@ -1,0 +1,44 @@
+
+.post {
+  border-bottom: double 5px #6c584c6e;
+  width: 90%;
+  padding-top: 20px;
+  padding-bottom: 10px;
+}
+.post .star {
+  color: orange;
+}
+.post .username {
+  text-decoration: none;
+  color: #5b4a40;
+  font-size: 20px;
+  padding-bottom: 5px;
+}
+.post .location {
+  text-decoration: none;
+  color: #877062;
+}
+.post .locationContainer {
+  display: flex;
+  gap: 5px;
+  align-items: center;
+  padding-top: 5px;
+  padding-bottom: 5px;
+}
+.post .image {
+  width: 100%;
+  object-fit: cover;
+}
+.post .topPost {
+  display: flex;
+  flex-direction: column;
+}
+@media screen and (min-width: 556px) {
+  .post{
+    width: 500px;
+  }
+}
+
+
+
+
