+++ src/components/Pages/Profile.jsx
@@ -3,11 +3,12 @@ import { useEffect, useState } from "react"
 import axios from "axios"
 import { useSelector } from "react-redux"
 import { selectToken, selectUserId } from "../../redux/slices/authSlice"
-import { useParams, useNavigate } from "react-router-dom"
+import { useParams } from "react-router-dom"
+import IndividualPost from "./IndividualPost/IndividualPost"
+import styles from "./Profile.module.css"
 
 function Profile() {
   const params = useParams()
-  const navigate = useNavigate()
   const userId = useSelector(selectUserId)
   const token = useSelector(selectToken)
   const [profile, setProfile] = useState({user: {}, posts: []})
@@ -99,33 +100,12 @@ function Profile() {
           </div>
         )) : null}
       </div>
-      <div>
+      <div className={styles.postContainer}>
         <h2>My Posts</h2>
         {profile.posts
           .sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt))
           .map((post) => (
-            <div key={post.id} onClick={() => {navigate(`/posts/${post.id}`)}}>
-              <h3>{post.title}</h3>
-              <p> 
-                {new Date(post.createdAt)
-                  .toDateString()
-                  .split(" ")
-                  .splice(1, 3)
-                  .join(" ")}
-              </p>
-              {/* add location address */}
-              <p>likes: {post.likes.length}</p>
-              <p>{post.content}</p>
-              <img src={post.url} alt="img" />
-              <p>Rating: {post.locationRating}</p>
-              {profile.id === userId && <button
-                onClick={() => {
-                  deleteHandler(post.id)
-                }}
-              >
-                Delete Post
-              </button>}
-            </div>
+            <IndividualPost post={post} style="post"/>
           ))}
       </div>
     </>
