+++ src/components/Pages/Profile.module.css
@@ -1,0 +1,6 @@
+.postContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
